feat(mcp): Stage 136 — add intake preview wrappers
Move the intake preview helper closure (9 pure helpers: intake + 4 per-
type + acceptance-map/stage-plan/agent-run-plan/evidence-plan) and their
tests into @conclave-ai/workspace-preview. Dashboard helper files become
thin re-export wrappers (export * from "@conclave-ai/workspace-preview/X")
so all existing imports keep working with zero call-site churn. Added
subpath exports + workspace dep in dashboard and mcp-workspace.

MCP Basic wrappers (packages/mcp-workspace/src/mcp-basic-preview-tools.mjs):
previewAcceptanceMap/StagePlan/AgentRunPlan/EvidencePlan — pure, read-only,
each returns {ok,kind,preview} + boundary (mutatesState/usesHostedExecution/
requiresPayment false, derivedPreviewOnly true); missing/invalid input
returns a safe error object (no throw). No server runtime wiring, no
network/env/mutation/LLM/payment/agent/PR-post.

One source of truth — no logic duplication. workspace-preview 93/93,
mcp-workspace 29/29, dashboard 300/300 (86 helper tests moved to package),
monorepo typecheck 57/57, dashboard build green.
