+++ apps/dashboard/package.json
@@ -11,6 +11,7 @@
     "test": "node --test test/*.test.mjs"
   },
   "dependencies": {
+    "@conclave-ai/workspace-preview": "workspace:*",
     "jszip": "^3.10.1",
     "next": "15.5.16",
     "react": "19.2.6",
