+++ apps/dashboard/src/lib/intake-prd.d.mts
@@ -1,13 +1,2 @@
-// Type declarations for intake-prd.mjs (Stage 102 — PRD/spec intake preview).
-
-export type PrdIntakePreview = {
-  productIntent: string;
-  likelyUsers: string[];
-  candidateUserFlows: string[];
-  candidateAcceptanceItems: string[];
-  missingQuestions: string[];
-  confidence: "low" | "medium" | "high";
-};
-
-export function buildPrdIntakePreview(rawInput: string): PrdIntakePreview;
-export const SAMPLE_PRD: string;
+// Stage 136 — compatibility wrapper types. See @conclave-ai/workspace-preview.
+export * from "@conclave-ai/workspace-preview/intake-prd";
