+++ apps/dashboard/src/lib/intake-url.d.mts
@@ -1,26 +1,2 @@
-// Type declarations for intake-url.mjs (Stage 103 — Product URL intake preview).
-
-export type ProductUrlPathType =
-  | "homepage"
-  | "pricing"
-  | "docs"
-  | "app"
-  | "demo"
-  | "blog"
-  | "unknown";
-
-export type ProductUrlIntakePreview = {
-  normalizedUrl: string;
-  domain: string;
-  pathType: ProductUrlPathType;
-  likelySurface: string;
-  reviewFocusAreas: string[];
-  candidateAcceptanceItems: string[];
-  missingQuestions: string[];
-  confidence: "low" | "medium" | "high";
-};
-
-export function buildProductUrlIntakePreview(
-  rawInput: string,
-): ProductUrlIntakePreview;
-export const SAMPLE_PRODUCT_URL: string;
+// Stage 136 — compatibility wrapper types. See @conclave-ai/workspace-preview.
+export * from "@conclave-ai/workspace-preview/intake-url";
