+++ packages/mcp-workspace/package.json
@@ -30,6 +30,7 @@
     "clean": "rm -rf dist .tsbuildinfo"
   },
   "dependencies": {
+    "@conclave-ai/workspace-preview": "workspace:*",
     "@modelcontextprotocol/sdk": "^1.29.0",
     "zod": "^3.24.0"
   },
