+++ packages/workspace-preview/package.json
@@ -15,6 +15,42 @@
     "./safety": {
       "types": "./src/safety.d.mts",
       "import": "./src/safety.mjs"
+    },
+    "./intake": {
+      "types": "./src/intake.d.mts",
+      "import": "./src/intake.mjs"
+    },
+    "./intake-prd": {
+      "types": "./src/intake-prd.d.mts",
+      "import": "./src/intake-prd.mjs"
+    },
+    "./intake-url": {
+      "types": "./src/intake-url.d.mts",
+      "import": "./src/intake-url.mjs"
+    },
+    "./intake-github-repo": {
+      "types": "./src/intake-github-repo.d.mts",
+      "import": "./src/intake-github-repo.mjs"
+    },
+    "./intake-ai-built-app": {
+      "types": "./src/intake-ai-built-app.d.mts",
+      "import": "./src/intake-ai-built-app.mjs"
+    },
+    "./intake-acceptance-map": {
+      "types": "./src/intake-acceptance-map.d.mts",
+      "import": "./src/intake-acceptance-map.mjs"
+    },
+    "./intake-stage-plan": {
+      "types": "./src/intake-stage-plan.d.mts",
+      "import": "./src/intake-stage-plan.mjs"
+    },
+    "./intake-agent-run-plan": {
+      "types": "./src/intake-agent-run-plan.d.mts",
+      "import": "./src/intake-agent-run-plan.mjs"
+    },
+    "./intake-evidence-plan": {
+      "types": "./src/intake-evidence-plan.d.mts",
+      "import": "./src/intake-evidence-plan.mjs"
     }
   },
   "files": [
