+++ packages/workspace-preview/src/index.d.mts
@@ -1,6 +1,16 @@
-// Public type entry for @conclave-ai/workspace-preview (Stage 135 skeleton).
+// Public type entry for @conclave-ai/workspace-preview.
 export {
   WORKSPACE_PREVIEW_PACKAGE,
   WORKSPACE_PREVIEW_SAFETY_RULES,
   getWorkspacePreviewSafetySummary,
 } from "./safety.d.mts";
+
+export * from "./intake.d.mts";
+export * from "./intake-prd.d.mts";
+export * from "./intake-url.d.mts";
+export * from "./intake-github-repo.d.mts";
+export * from "./intake-ai-built-app.d.mts";
+export * from "./intake-acceptance-map.d.mts";
+export * from "./intake-stage-plan.d.mts";
+export * from "./intake-agent-run-plan.d.mts";
+export * from "./intake-evidence-plan.d.mts";
